+++ src/entities/chat/model/useMainChat.ts
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from 'react';
 
-import { useAuthStore } from '@/features/auth'; // TODO: 오류!! useAuthStore 위치 변경 해야함(shared)
+import { useAuthStore } from '@/entities/user';
 import { mockChannels, mockFriends } from '@/mocks/chat.mock';
 import { useSocket } from '@/shared/api/socket';
 import type { Friend } from '@/shared/model';
