+++ src/features/auth/index.ts
@@ -2,7 +2,6 @@ export { postLogin } from './api/postLogin';
 export { postRefreshToken } from './api/postRefreshToken';
 export { postSignup } from './api/postSignup';
 export * from './model/types';
-export { useAuthStore } from './model/useAuthStore';
 export { authQueries } from './queries/authQueries';
 export { LoginForm } from './ui/LoginForm';
 export { SignupForm } from './ui/SignupForm';
