+++ src/features/lobby/model/useRoom.ts
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { useParams } from 'react-router';
 
 import type { RoomState } from '@/entities/game';
-import { useAuthStore } from '@/features/auth'; // TODO: 오류. 다른 features 참고 불가능함
+import { useAuthStore } from '@/entities/user';
 import { useSocket } from '@/shared/api/socket';
 import {
   selectCanStartGame,
