+++ src/pages/WelcomePage.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router';
 import { motion } from 'framer-motion';
 
-import { useAuthStore } from '@/features/auth'; // TODO: useAuthStore 경로변경
+import { useAuthStore } from '@/entities/user';
 
 const WelcomePage = () => {
   const navigate = useNavigate();
